feat(atomic): slight rename
