+++ packages/atomic/src/components/insight/atomic-insight-tab/atomic-insight-tab.tsx
@@ -8,7 +8,7 @@ import {
 } from '../../../utils/initialization-utils';
 import {randomID} from '../../../utils/utils';
 import {Button} from '../../common/button';
-import {dispatchTabLoaded, TabCommon} from '../../common/tabs/tabs-common';
+import {dispatchTabLoaded, TabCommon} from '../../common/tabs/tab-common';
 import {InsightBindings} from '../atomic-insight-interface/atomic-insight-interface';
 
 /**
